Rename existingThoughtChange -> editThought
