+++ src/action-creators/index.ts
@@ -34,7 +34,7 @@ export const editingValue = reducerToThunk<typeof reducers.editingValue>('editin
 export const error = reducerToThunk<typeof reducers.error>('error')
 export const expandHoverTop = reducerToThunk<typeof reducers.expandHoverTop>('expandHoverTop')
 export const expandBottom = reducerToThunk<typeof reducers.expandBottom>('expandBottom')
-export const existingThoughtChange = reducerToThunk<typeof reducers.existingThoughtChange>('existingThoughtChange')
+export const editThought = reducerToThunk<typeof reducers.editThought>('editThought')
 export const existingThoughtDelete = reducerToThunk<typeof reducers.existingThoughtDelete>('existingThoughtDelete')
 export const existingThoughtMove = reducerToThunk<typeof reducers.existingThoughtMove>('existingThoughtMove')
 export const extractThought = reducerToThunk<typeof reducers.extractThought>('extractThought')
