+++ src/reducers/index.ts
@@ -28,7 +28,7 @@ export { default as editingValue } from './editingValue'
 export { default as error } from './error'
 export { default as expandHoverTop } from './expandHoverTop'
 export { default as expandBottom } from './expandBottom'
-export { default as existingThoughtChange } from './existingThoughtChange'
+export { default as editThought } from './editThought'
 export { default as existingThoughtDelete } from './existingThoughtDelete'
 export { default as existingThoughtMove } from './existingThoughtMove'
 export { default as expandContextThought } from './expandContextThought'
